refactor: login from routes to components
to allow for reusability if required
